src/requirements.txt
action-msgs
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
builtin-interfaces==1.2.1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromadb==0.5.18
click==8.1.7
colorama==0.4.6
composition-interfaces==1.2.1
cryptography==43.0.1
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
domain-coordinator==0.10.0
enum-compat==0.0.3
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
geometry-msgs==4.2.4
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-run==0.10.9
google-cloud-scheduler==2.13.5
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
intuit-oauth==1.2.6
itsdangerous==2.2.0
Jinja2==3.1.4
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.1
message-filters==4.3.5
msgspec==0.18.6
multidict==6.1.0
nav-msgs==4.2.4
oauthlib==3.2.2
osrf-pycommon==2.0.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
py-trello==0.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==1.1.2
pytz==2024.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
requests==2.32.3
requests-oauthlib==2.0.0
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
scheduler==0.8.7
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
slack_sdk==3.33.1
slackclient==2.9.4
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
topic-monitor==0.20.5
tqdm==4.66.5
trajectory-msgs==4.2.4
turtlesim==1.4.2
typeguard==4.3.0
typing_extensions==4.12.2
unique-identifier-msgs==2.2.1
uritemplate==4.1.1
urllib3==2.2.3
visualization-msgs==4.2.4
Werkzeug==3.0.4
yarl==1.15.3
